requirements.txt
gunicorn
azure-ai-inference
python-dotenv
flask
bs4
pandas
requests
transformers
torch
psycopg
pytz
